+++ compiler/generator/soul/soul_instructions.hh
@@ -302,7 +302,7 @@ class SOULInstVisitor : public TextInstVisitor {
     virtual void visit(FloatArrayNumInst* inst)
     {
         char sep = '(';
-        for (unsigned int i = 0; i < inst->fNumTable.size(); i++) {
+        for (size_t i = 0; i < inst->fNumTable.size(); i++) {
             *fOut << sep << checkFloat(inst->fNumTable[i]);
             sep = ',';
         }
@@ -312,7 +312,7 @@ class SOULInstVisitor : public TextInstVisitor {
     virtual void visit(Int32ArrayNumInst* inst)
     {
         char sep = '(';
-        for (unsigned int i = 0; i < inst->fNumTable.size(); i++) {
+        for (size_t i = 0; i < inst->fNumTable.size(); i++) {
             *fOut << sep << inst->fNumTable[i];
             sep = ',';
         }
@@ -322,7 +322,7 @@ class SOULInstVisitor : public TextInstVisitor {
     virtual void visit(DoubleArrayNumInst* inst)
     {
         char sep = '(';
-        for (unsigned int i = 0; i < inst->fNumTable.size(); i++) {
+        for (size_t i = 0; i < inst->fNumTable.size(); i++) {
             *fOut << sep << checkDouble(inst->fNumTable[i]);
             sep = ',';
         }
