+++ compiler/generator/wasm/wast_instructions.hh
@@ -49,7 +49,7 @@ class WASTInstVisitor : public TextInstVisitor, public WASInst {
     {
         bool dot   = false;
         int  e_pos = -1;
-        for (unsigned int i = 0; i < str.size(); i++) {
+        for (size_t i = 0; i < str.size(); i++) {
             if (str[i] == '.') {
                 dot = true;
                 break;
